mcp_servers/wikipedia-mcp/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "wikipedia-mcp"
version = "1.5.4"
authors = [
    {name = "Ravi Kumar E", email = "ravikumar@ravikumar-dev.me"},
]
description = "MCP server for Wikipedia API"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
keywords = ["mcp", "wikipedia", "model-context-protocol", "ai", "llm"]
dependencies = [
    "fastmcp>=2.3.0",
    "wikipedia-api>=0.8.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/rudra-ravi/wikipedia-mcp"
Repository = "https://github.com/rudra-ravi/wikipedia-mcp"
Issues = "https://github.com/rudra-ravi/wikipedia-mcp/issues"
Documentation = "https://github.com/rudra-ravi/wikipedia-mcp#readme"

[project.scripts]
wikipedia-mcp = "wikipedia_mcp.__main__:main"

[tool.setuptools]
packages.find.where = ["."]
packages.find.include = ["wikipedia_mcp*"]
package-data.wikipedia_mcp = ["py.typed"]
license-files = ["LICENSE"] 
